Medical_RAGAS/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
openai==1.93.1
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
faiss-cpu==1.11.0
pdfplumber==0.11.7
tiktoken==0.9.0
python-dotenv==1.0.1
ragas==0.1.3
datasets==3.6.0
evaluate==0.4.4
sentence-transformers==2.7.0 